+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+
+- Starr app Media Management API configuration ([#221](https://github.com/kiriwalawren/nixflix/pull/221)).
+
 ## [1.2.0] - 2026-05-20
 
 ### Added